Déplace handleAjaxError vers calendar/utils.js
